Adaptive Workflow mostly works
