+++ examples/workflows/workflow_adaptive/graded_report.md
@@ -1,0 +1,37 @@
+### Graded Report: The Battle of Glimmerwood
+
+#### Title Analysis:
+- **Title**: The title "The Battle of Glimmerwood" sets the tone for a story that focuses on conflict within a mystical forest setting. It effectively hints at a central conflict integral to the plot.
+
+#### Setting:
+- **Location**: The story is set in Glimmerwood, a mystical forest known for its radiant trees and magical creatures. 
+- **World-building**: The description of Glimmerwood and its inhabitants, particularly the Glimmerfoxes, creates a vivid and magical world.
+- **Effectiveness**: The setting is well-integrated into the storyline, enhancing the magical and mythical aspects of the narrative.
+
+#### Plot and Structure:
+- **Introduction**: The story begins by establishing a peaceful village within Glimmerwood, sharing space with magical creatures.
+- **Conflict**: The peace is disrupted by the Dark Marauders, led by Captain Thorn, seeking Glimmerstones rumored to grant immortality.
+- **Resolution**: The protagonist, Elara, devises a plan to defend the village with the help of the Glimmerfoxes, leading to the capture of the marauders.
+- **Cohesiveness**: The plot flows logically from the introduction of the setting to the resolution of the conflict, with each element serving the overarching narrative.
+
+#### Characterization:
+- **Protagonist**: Elara, a courageous young girl, is portrayed as the hero who saves the village.
+- **Antagonists**: The Dark Marauders, particularly Captain Thorn, serve as effective antagonists driving the conflict.
+- **Character Development**: Elara’s leadership and bravery are emphasized, reinforcing her role as the "Guardian of Glimmerwood."
+
+#### Themes:
+- **Good vs. Evil**: A classic theme explored through the conflict between the villagers and the marauders.
+- **Courage and Leadership**: Embodied by Elara’s actions in rallying the villagers and forming a successful defense.
+- **Mystery and Intrigue**: The story leaves room for doubt regarding the true power of the Glimmerstones and potential underlying motives.
+
+#### Style and Language:
+- **Descriptive Language**: The story utilizes vivid descriptions to bring the mystical elements of Glimmerwood to life.
+- **Grammar and Syntax**: There are minor grammatical errors (e.g., "knowed" instead of "known," "was" instead of "were," "shaterred" instead of "shattered," etc.) that slightly detract from the reading experience.
+- **Pacing**: The story maintains a smooth pace, keeping readers engaged from start to finish.
+
+#### Overall Impression:
+- **Strengths**: Strong world-building and a well-developed plot with a clear conflict and resolution. The central character is well-developed, and her heroism is inspiring.
+- **Areas for Improvement**: Attention to grammatical detail and sentence structure would enhance clarity and polish.
+
+#### Grade:
+- Based on the elements examined, the story merits a grade of **B+**. The creative ideas and compelling narrative are slightly hindered by grammatical inconsistencies, which, if corrected, could elevate the story further.
