+++ tests/workflows/adaptive/test_adaptive_workflow_edge_cases.py
@@ -5,20 +5,20 @@ Edge case and error handling tests for Adaptive Workflow V2
 import pytest
 import asyncio
 from unittest.mock import AsyncMock, MagicMock, patch
-from datetime import datetime, timedelta
-from typing import Optional, Any
+from datetime import datetime
+from typing import Optional
 
-from mcp_agent.workflows.adaptive.adaptive_workflow_v2 import AdaptiveWorkflowV2
-from mcp_agent.workflows.adaptive.models_v2 import (
+from mcp_agent.workflows.adaptive.adaptive_workflow import AdaptiveWorkflow
+from mcp_agent.workflows.adaptive.models import (
     TaskType,
     ResearchAspect,
     SubagentResult,
     SynthesisDecision,
     ExecutionMemory,
 )
-from mcp_agent.workflows.adaptive.memory_v2 import MemoryManager
+from mcp_agent.workflows.adaptive.memory import MemoryManager
 from mcp_agent.agents.agent import Agent
-from mcp_agent.workflows.llm.augmented_llm import AugmentedLLM, RequestParams
+from mcp_agent.workflows.llm.augmented_llm import AugmentedLLM
 
 
 class MockAugmentedLLM(AugmentedLLM):
@@ -38,8 +38,10 @@ class MockAugmentedLLM(AugmentedLLM):
         return await self.generate_str_mock(message, request_params)
 
     async def generate_structured(self, message, response_model, request_params=None):
-        return await self.generate_structured_mock(message, response_model, request_params)
-    
+        return await self.generate_structured_mock(
+            message, response_model, request_params
+        )
+
     def message_str(self, message, content_only=False):
         return self.message_str_mock(message, content_only)
 
@@ -50,29 +52,29 @@ class TestErrorHandling:
     @pytest.mark.asyncio
     async def test_subagent_execution_failure(self, mock_llm_factory, mock_context):
         """Test handling of subagent execution failures"""
-        workflow = AdaptiveWorkflowV2(
-            llm_factory=mock_llm_factory,
-            context=mock_context
-        )
-        
+        workflow = AdaptiveWorkflow(llm_factory=mock_llm_factory, context=mock_context)
+
         # Set up workflow state
         workflow._current_memory = ExecutionMemory(
             execution_id="test",
             objective="Test failure handling",
-            task_type=TaskType.RESEARCH
+            task_type=TaskType.RESEARCH,
         )
-        
+
         # Create aspect
         aspect = ResearchAspect(
             name="Failing Aspect",
             objective="This will fail",
-            required_servers=["non_existent_server"]
+            required_servers=["non_existent_server"],
         )
-        
+
         # Mock the agent creation to raise error
-        with patch('mcp_agent.agents.agent.Agent.__aenter__', side_effect=Exception("Server connection failed")):
+        with patch(
+            "mcp_agent.agents.agent.Agent.__aenter__",
+            side_effect=Exception("Server connection failed"),
+        ):
             result = await workflow._execute_single_aspect(aspect, None, MagicMock())
-        
+
         assert result.success is False
         assert result.error == "Server connection failed"
         assert result.findings is None
@@ -80,36 +82,35 @@ class TestErrorHandling:
     @pytest.mark.asyncio
     async def test_memory_save_failure(self, mock_llm_factory, mock_context):
         """Test handling of memory save failures"""
+
         # Create memory manager with failing backend
         class FailingBackend:
             async def save(self, *args):
                 raise IOError("Disk full")
-            
+
             async def load(self, *args):
                 return None
-            
+
             async def delete(self, *args):
                 pass
-            
+
             async def list_executions(self):
                 return {}
-        
+
         memory_manager = MemoryManager(backend=FailingBackend())
-        
-        workflow = AdaptiveWorkflowV2(
+
+        workflow = AdaptiveWorkflow(
             llm_factory=mock_llm_factory,
             memory_manager=memory_manager,
-            context=mock_context
+            context=mock_context,
         )
-        
+
         # The workflow should handle memory save failures gracefully
         # and continue execution
         workflow._current_memory = ExecutionMemory(
-            execution_id="test",
-            objective="Test",
-            task_type=TaskType.RESEARCH
+            execution_id="test", objective="Test", task_type=TaskType.RESEARCH
         )
-        
+
         # This should not raise an exception
         try:
             await workflow.memory_manager.save_memory(workflow._current_memory)
@@ -120,31 +121,26 @@ class TestErrorHandling:
     @pytest.mark.asyncio
     async def test_llm_exception_handling(self, mock_llm_factory, mock_context):
         """Test handling of exceptions during LLM execution"""
-        workflow = AdaptiveWorkflowV2(
-            llm_factory=mock_llm_factory,
-            context=mock_context
-        )
-        
+        workflow = AdaptiveWorkflow(llm_factory=mock_llm_factory, context=mock_context)
+
         # Set up workflow state
         workflow._current_memory = ExecutionMemory(
-            execution_id="test",
-            objective="Test exception",
-            task_type=TaskType.RESEARCH
+            execution_id="test", objective="Test exception", task_type=TaskType.RESEARCH
         )
-        
+
         # Create aspect that will fail
         aspect = ResearchAspect(
-            name="Exception Test",
-            objective="This will throw an exception"
+            name="Exception Test", objective="This will throw an exception"
         )
-        
+
         # Mock the execute_single_aspect to raise an exception
         original_execute = workflow._execute_single_aspect
+
         async def failing_execute(*args, **kwargs):
             raise RuntimeError("Simulated LLM failure")
-        
+
         workflow._execute_single_aspect = failing_execute
-        
+
         # Execute should handle the exception gracefully
         try:
             result = await workflow._execute_single_aspect(aspect, None, MagicMock())
@@ -163,26 +159,24 @@ class TestBoundaryConditions:
         # Mock analysis
         mock_analysis = MagicMock()
         mock_analysis.task_type = TaskType.RESEARCH
-        
+
         # Create mock LLM instance
         mock_llm_instance = MockAugmentedLLM()
         mock_llm_instance.generate_structured_mock.return_value = mock_analysis
         mock_llm_instance.generate_mock.return_value = ["Empty report - no iterations"]
-        
+
         # Make factory return our configured mock
         mock_llm_factory.side_effect = lambda agent: mock_llm_instance
-        
-        workflow = AdaptiveWorkflowV2(
-            llm_factory=mock_llm_factory,
-            context=mock_context,
-            max_iterations=0
+
+        workflow = AdaptiveWorkflow(
+            llm_factory=mock_llm_factory, context=mock_context, max_iterations=0
         )
-        
+
         # Mock basic setup
         workflow._extract_objective = AsyncMock(return_value="Test")
-        
+
         result = await workflow.generate("Test objective")
-        
+
         # Should complete immediately without any research iterations
         assert result == ["Empty report - no iterations"]
 
@@ -192,28 +186,25 @@ class TestBoundaryConditions:
         # Mock empty plan
         mock_plan = MagicMock()
         mock_plan.aspects = []  # No aspects to research
-        
+
         # Create mock LLM instance
         mock_llm_instance = MockAugmentedLLM()
         mock_llm_instance.generate_structured_mock.return_value = mock_plan
-        
+
         # Make factory return our configured mock
         mock_llm_factory.side_effect = lambda agent: mock_llm_instance
-        
-        workflow = AdaptiveWorkflowV2(
-            llm_factory=mock_llm_factory,
-            context=mock_context
-        )
-        
+
+        workflow = AdaptiveWorkflow(llm_factory=mock_llm_factory, context=mock_context)
+
         # Set up workflow state
         workflow._current_memory = ExecutionMemory(
             execution_id="test",
             objective="Test empty aspects",
-            task_type=TaskType.RESEARCH
+            task_type=TaskType.RESEARCH,
         )
-        
+
         aspects = await workflow._plan_research(MagicMock())
-        
+
         assert aspects == []
 
     @pytest.mark.asyncio
@@ -222,33 +213,30 @@ class TestBoundaryConditions:
         # Create mock LLM instance
         mock_llm_instance = MockAugmentedLLM()
         mock_llm_instance.generate_mock.return_value = ["No successful research"]
-        
+
         # Make factory return our configured mock
         mock_llm_factory.side_effect = lambda agent: mock_llm_instance
-        
-        workflow = AdaptiveWorkflowV2(
-            llm_factory=mock_llm_factory,
-            context=mock_context
-        )
-        
+
+        workflow = AdaptiveWorkflow(llm_factory=mock_llm_factory, context=mock_context)
+
         # Create failing results
         failed_results = [
             SubagentResult(
                 aspect_name="Fail 1",
                 success=False,
                 error="Error 1",
-                start_time=datetime.now()
+                start_time=datetime.now(),
             ),
             SubagentResult(
                 aspect_name="Fail 2",
                 success=False,
                 error="Error 2",
-                start_time=datetime.now()
-            )
+                start_time=datetime.now(),
+            ),
         ]
-        
+
         synthesis = await workflow._synthesize_results(failed_results, MagicMock())
-        
+
         # Should still return something, even with all failures
         assert synthesis == ["No successful research"]
 
@@ -259,15 +247,13 @@ class TestConcurrencyAndRaceConditions:
     @pytest.mark.asyncio
     async def test_parallel_subagent_execution(self, mock_llm_factory, mock_context):
         """Test parallel execution of multiple subagents"""
-        workflow = AdaptiveWorkflowV2(
-            llm_factory=mock_llm_factory,
-            context=mock_context,
-            enable_parallel=True
+        workflow = AdaptiveWorkflow(
+            llm_factory=mock_llm_factory, context=mock_context, enable_parallel=True
         )
-        
+
         # Track execution order
         execution_order = []
-        
+
         async def mock_execute(aspect, params, span):
             execution_order.append(f"start_{aspect.name}")
             await asyncio.sleep(0.1)  # Simulate work
@@ -276,29 +262,27 @@ class TestConcurrencyAndRaceConditions:
                 aspect_name=aspect.name,
                 success=True,
                 findings=f"Result for {aspect.name}",
-                start_time=datetime.now()
+                start_time=datetime.now(),
             )
-        
+
         workflow._execute_single_aspect = mock_execute
         workflow._current_memory = ExecutionMemory(
-            execution_id="test",
-            objective="Test parallel",
-            task_type=TaskType.RESEARCH
+            execution_id="test", objective="Test parallel", task_type=TaskType.RESEARCH
         )
-        
+
         # Create multiple aspects
         aspects = [
             ResearchAspect(name=f"Aspect_{i}", objective=f"Research {i}")
             for i in range(3)
         ]
-        
+
         results = await workflow._execute_research(aspects, None, MagicMock())
-        
+
         # All should start before any finish (parallel execution)
         assert "start_Aspect_0" in execution_order
         assert "start_Aspect_1" in execution_order
         assert "start_Aspect_2" in execution_order
-        
+
         # Verify all completed
         assert len(results) == 3
         assert all(r.success for r in results)
@@ -306,43 +290,42 @@ class TestConcurrencyAndRaceConditions:
     @pytest.mark.asyncio
     async def test_sequential_execution(self, mock_llm_factory, mock_context):
         """Test sequential execution when parallel is disabled"""
-        workflow = AdaptiveWorkflowV2(
-            llm_factory=mock_llm_factory,
-            context=mock_context,
-            enable_parallel=False
+        workflow = AdaptiveWorkflow(
+            llm_factory=mock_llm_factory, context=mock_context, enable_parallel=False
         )
-        
+
         execution_order = []
-        
+
         async def mock_execute(aspect, params, span):
             execution_order.append(f"start_{aspect.name}")
             await asyncio.sleep(0.05)
             execution_order.append(f"end_{aspect.name}")
             return SubagentResult(
-                aspect_name=aspect.name,
-                success=True,
-                start_time=datetime.now()
+                aspect_name=aspect.name, success=True, start_time=datetime.now()
             )
-        
+
         workflow._execute_single_aspect = mock_execute
         workflow._current_memory = ExecutionMemory(
             execution_id="test",
             objective="Test sequential",
-            task_type=TaskType.RESEARCH
+            task_type=TaskType.RESEARCH,
         )
-        
+
         aspects = [
             ResearchAspect(name=f"Aspect_{i}", objective=f"Research {i}")
             for i in range(3)
         ]
-        
+
         await workflow._execute_research(aspects, None, MagicMock())
-        
+
         # Should complete one before starting next (sequential)
         expected_order = [
-            "start_Aspect_0", "end_Aspect_0",
-            "start_Aspect_1", "end_Aspect_1",
-            "start_Aspect_2", "end_Aspect_2"
+            "start_Aspect_0",
+            "end_Aspect_0",
+            "start_Aspect_1",
+            "end_Aspect_1",
+            "start_Aspect_2",
+            "end_Aspect_2",
         ]
         assert execution_order == expected_order
 
@@ -353,104 +336,105 @@ class TestResourceManagement:
     @pytest.mark.asyncio
     async def test_cost_budget_exceeded(self, mock_llm_factory, mock_context):
         """Test stopping when cost budget is exceeded"""
-        workflow = AdaptiveWorkflowV2(
+        workflow = AdaptiveWorkflow(
             llm_factory=mock_llm_factory,
             context=mock_context,
             cost_budget=2.0,  # Low budget
-            max_iterations=10
+            max_iterations=10,
         )
-        
+
         # Set up memory with high cost
         workflow._current_memory = ExecutionMemory(
             execution_id="test",
             objective="Test cost limit",
-            total_cost=1.9  # Almost at budget
+            total_cost=1.9,  # Almost at budget
         )
-        
+
         # Add a result that exceeds budget
         workflow._current_memory.subagent_results.append(
             SubagentResult(
                 aspect_name="Expensive",
                 success=True,
                 cost=0.5,  # This will exceed budget
-                start_time=datetime.now()
+                start_time=datetime.now(),
             )
         )
         workflow._current_memory.total_cost = 2.4
-        
+
         # Check should_continue
         assert workflow._should_continue(0) is False
 
     @pytest.mark.asyncio
     async def test_memory_cleanup_on_exception(self, mock_llm_factory, mock_context):
         """Test that memory is properly cleaned up on exceptions"""
-        workflow = AdaptiveWorkflowV2(
-            llm_factory=mock_llm_factory,
-            context=mock_context
-        )
-        
+        workflow = AdaptiveWorkflow(llm_factory=mock_llm_factory, context=mock_context)
+
         # Mock to raise exception during execution
         workflow._execute_workflow = AsyncMock(
             side_effect=Exception("Unexpected error")
         )
-        
+
         with pytest.raises(Exception) as exc_info:
             await workflow.generate("Test objective")
-        
+
         assert str(exc_info.value) == "Unexpected error"
-        
+
         # Memory should be reset
-        assert workflow._current_execution_id is None or workflow._current_memory is None
+        assert (
+            workflow._current_execution_id is None or workflow._current_memory is None
+        )
 
 
 class TestComplexScenarios:
     """Tests for complex real-world scenarios"""
 
     @pytest.mark.asyncio
-    async def test_mixed_predefined_and_new_agents(self, mock_llm_factory, mock_context):
+    async def test_mixed_predefined_and_new_agents(
+        self, mock_llm_factory, mock_context
+    ):
         """Test workflow using both predefined and newly created agents"""
         # Create predefined agents
         web_agent = Agent(name="WebAgent", instruction="Web search specialist")
         code_agent = Agent(name="CodeAgent", instruction="Code analysis specialist")
-        
-        workflow = AdaptiveWorkflowV2(
+
+        workflow = AdaptiveWorkflow(
             llm_factory=mock_llm_factory,
             available_agents=[web_agent, code_agent],
             available_servers=["web_search", "filesystem", "database"],
-            context=mock_context
+            context=mock_context,
         )
-        
+
         # Set up aspects using mixed agents
         aspects = [
             ResearchAspect(
                 name="Web Research",
                 objective="Search online",
-                use_predefined_agent="WebAgent"
+                use_predefined_agent="WebAgent",
             ),
             ResearchAspect(
                 name="Code Analysis",
                 objective="Analyze code",
-                use_predefined_agent="CodeAgent"
+                use_predefined_agent="CodeAgent",
             ),
             ResearchAspect(
                 name="Database Research",
                 objective="Query database",
-                required_servers=["database"]  # Will create new agent
-            )
+                required_servers=["database"],  # Will create new agent
+            ),
         ]
-        
+
         workflow._current_memory = ExecutionMemory(
             execution_id="test",
             objective="Mixed agents test",
-            task_type=TaskType.RESEARCH
+            task_type=TaskType.RESEARCH,
         )
-        
+
         # Mock LLM responses
         mock_llm = mock_llm_factory.return_value
         mock_llm.generate_str_mock.return_value = "Research complete"
-        
+
         results = await workflow._execute_research(aspects, None, MagicMock())
-        
+
         assert len(results) == 3
         # Verify the factory was called for new agents but not predefined
         assert mock_llm_factory.call_count >= 1  # At least for the new agent
@@ -458,61 +442,53 @@ class TestComplexScenarios:
     @pytest.mark.asyncio
     async def test_adaptive_replanning(self, mock_llm_factory, mock_context):
         """Test adaptive replanning based on findings"""
-        workflow = AdaptiveWorkflowV2(
-            llm_factory=mock_llm_factory,
-            context=mock_context,
-            max_iterations=3
+        workflow = AdaptiveWorkflow(
+            llm_factory=mock_llm_factory, context=mock_context, max_iterations=3
         )
-        
+
         # Track planning calls
         plan_calls = []
-        
+
         async def mock_plan(span):
             iteration = len(plan_calls)
             plan_calls.append(iteration)
-            
+
             if iteration == 0:
                 # First plan
-                return [
-                    ResearchAspect(name="Initial", objective="Start research")
-                ]
+                return [ResearchAspect(name="Initial", objective="Start research")]
             elif iteration == 1:
                 # Adaptive replan based on findings
-                return [
-                    ResearchAspect(name="Followup", objective="Dig deeper")
-                ]
+                return [ResearchAspect(name="Followup", objective="Dig deeper")]
             else:
                 # No more aspects
                 return []
-        
+
         workflow._plan_research = mock_plan
-        
+
         # Mock analyze_objective to return a proper TaskType
         workflow._analyze_objective = AsyncMock(return_value=TaskType.RESEARCH)
-        
+
         # Mock other components
         workflow._current_memory = ExecutionMemory(
-            execution_id="test",
-            objective="Adaptive test",
-            task_type=TaskType.RESEARCH
+            execution_id="test", objective="Adaptive test", task_type=TaskType.RESEARCH
         )
-        
+
         # Configure mock LLM instance properly
         mock_llm_instance = MockAugmentedLLM()
         mock_llm_instance.generate_str_mock.return_value = "Finding"
         mock_llm_instance.generate_mock.return_value = ["Synthesis"]
         mock_llm_instance.message_str_mock.return_value = "Synthesis string"
-        
+
         # Mock decision to continue first, then complete
         decisions = [
             SynthesisDecision(is_complete=False, confidence=0.5, reasoning="Need more"),
-            SynthesisDecision(is_complete=True, confidence=0.9, reasoning="Complete")
+            SynthesisDecision(is_complete=True, confidence=0.9, reasoning="Complete"),
         ]
         mock_llm_instance.generate_structured_mock.side_effect = decisions
-        
+
         # Make factory return our configured mock
         mock_llm_factory.side_effect = lambda agent: mock_llm_instance
-        
+
         # Mock the subagent execution to avoid the keys() error
         async def mock_execute_single(aspect, params, span):
             return SubagentResult(
@@ -521,14 +497,14 @@ class TestComplexScenarios:
                 findings=f"Findings for {aspect.name}",
                 start_time=datetime.now(),
                 end_time=datetime.now(),
-                cost=0.1
+                cost=0.1,
             )
-        
+
         workflow._execute_single_aspect = mock_execute_single
-        
+
         # Execute partial workflow
         await workflow._execute_workflow("Test", None, MagicMock())
-        
+
         # Should have planned twice
         assert len(plan_calls) == 2
         assert plan_calls == [0, 1]
@@ -538,7 +514,7 @@ class TestComplexScenarios:
 def mock_context():
     """Mock context for edge case tests"""
     from mcp_agent.core.context import Context
-    
+
     context = MagicMock(spec=Context)
     context.server_registry = MagicMock()
     context.executor = MagicMock()
@@ -554,9 +530,10 @@ def mock_context():
 @pytest.fixture
 def mock_llm_factory():
     """Mock LLM factory for edge case tests"""
+
     def factory(agent):
         return MockAugmentedLLM(agent=agent)
-    
+
     mock_factory = MagicMock(side_effect=factory)
     mock_factory.return_value = MockAugmentedLLM()
     return mock_factory
