+++ tests/workflows/adaptive/test_memory.py
@@ -3,19 +3,15 @@ Tests for Adaptive Workflow V2 Memory Management
 """
 
 import pytest
-import asyncio
-import json
-from pathlib import Path
 from datetime import datetime
-from typing import Dict, Any
 
-from mcp_agent.workflows.adaptive.memory_v2 import (
+from mcp_agent.workflows.adaptive.memory import (
     MemoryManager,
     InMemoryBackend,
     FileSystemBackend,
     AdaptiveMemory,
 )
-from mcp_agent.workflows.adaptive.models_v2 import (
+from mcp_agent.workflows.adaptive.models import (
     ExecutionMemory,
     TaskType,
     SubagentResult,
@@ -29,24 +25,24 @@ class TestInMemoryBackend:
     async def test_save_and_load(self):
         """Test basic save and load operations"""
         backend = InMemoryBackend()
-        
+
         memory = ExecutionMemory(
             execution_id="test-123",
             objective="Test objective",
             task_type=TaskType.RESEARCH,
             iterations=5,
-            total_cost=10.0
+            total_cost=10.0,
         )
-        
+
         # Save
         await backend.save("test-123", memory)
-        
+
         # Load
         loaded = await backend.load("test-123")
         assert loaded is not None
         assert loaded.execution_id == "test-123"
         assert loaded.iterations == 5
-        
+
         # Load non-existent
         assert await backend.load("non-existent") is None
 
@@ -54,16 +50,14 @@ class TestInMemoryBackend:
     async def test_delete(self):
         """Test deletion"""
         backend = InMemoryBackend()
-        
+
         memory = ExecutionMemory(
-            execution_id="test-456",
-            objective="Test",
-            task_type=TaskType.ACTION
+            execution_id="test-456", objective="Test", task_type=TaskType.ACTION
         )
-        
+
         await backend.save("test-456", memory)
         assert await backend.load("test-456") is not None
-        
+
         await backend.delete("test-456")
         assert await backend.load("test-456") is None
 
@@ -71,17 +65,17 @@ class TestInMemoryBackend:
     async def test_list_executions(self):
         """Test listing executions"""
         backend = InMemoryBackend()
-        
+
         # Save multiple executions
         for i in range(3):
             memory = ExecutionMemory(
                 execution_id=f"test-{i}",
                 objective=f"Objective {i}",
                 task_type=TaskType.RESEARCH,
-                iterations=i
+                iterations=i,
             )
             await backend.save(f"test-{i}", memory)
-        
+
         # List
         executions = await backend.list_executions()
         assert len(executions) == 3
@@ -92,20 +86,20 @@ class TestInMemoryBackend:
     async def test_pattern_storage(self):
         """Test pattern storage and retrieval"""
         backend = InMemoryBackend()
-        
+
         # Save patterns
         pattern1 = {"approach": "web_search", "success": True}
         pattern2 = {"approach": "code_analysis", "success": False}
-        
+
         await backend.save_pattern("research:pattern1", pattern1)
         await backend.save_pattern("research:pattern2", pattern2)
         await backend.save_pattern("action:pattern1", {"approach": "file_edit"})
-        
+
         # Load patterns by type
         research_patterns = await backend.load_patterns("research")
         assert len(research_patterns) == 2
         assert "research:pattern1" in research_patterns
-        
+
         action_patterns = await backend.load_patterns("action")
         assert len(action_patterns) == 1
 
@@ -117,21 +111,21 @@ class TestFileSystemBackend:
     async def test_save_and_load(self, tmp_path):
         """Test basic save and load with filesystem"""
         backend = FileSystemBackend(str(tmp_path / "test_memory"))
-        
+
         memory = ExecutionMemory(
             execution_id="fs-test-123",
             objective="Filesystem test",
             task_type=TaskType.HYBRID,
-            start_time=datetime.now()
+            start_time=datetime.now(),
         )
-        
+
         # Save
         await backend.save("fs-test-123", memory)
-        
+
         # Verify file exists
         file_path = tmp_path / "test_memory" / "executions" / "fs-test-123.json"
         assert file_path.exists()
-        
+
         # Load
         loaded = await backend.load("fs-test-123")
         assert loaded is not None
@@ -141,16 +135,16 @@ class TestFileSystemBackend:
     async def test_persistence_across_instances(self, tmp_path):
         """Test that data persists across backend instances"""
         path = str(tmp_path / "persistent_memory")
-        
+
         # First instance saves
         backend1 = FileSystemBackend(path)
         memory = ExecutionMemory(
             execution_id="persist-test",
             objective="Test persistence",
-            task_type=TaskType.ACTION
+            task_type=TaskType.ACTION,
         )
         await backend1.save("persist-test", memory)
-        
+
         # Second instance loads
         backend2 = FileSystemBackend(path)
         loaded = await backend2.load("persist-test")
@@ -161,16 +155,16 @@ class TestFileSystemBackend:
     async def test_pattern_persistence(self, tmp_path):
         """Test pattern storage in filesystem"""
         backend = FileSystemBackend(str(tmp_path / "pattern_test"))
-        
+
         # Save patterns
         patterns = {
             "pattern1": {"success": True, "tools": ["web_search"]},
-            "pattern2": {"success": False, "error": "timeout"}
+            "pattern2": {"success": False, "error": "timeout"},
         }
-        
+
         for key, data in patterns.items():
             await backend.save_pattern(f"research:{key}", data)
-        
+
         # Load patterns
         loaded = await backend.load_patterns("research")
         assert len(loaded) == 2
@@ -180,18 +174,18 @@ class TestFileSystemBackend:
     async def test_corrupted_file_handling(self, tmp_path):
         """Test handling of corrupted JSON files"""
         backend = FileSystemBackend(str(tmp_path / "corrupt_test"))
-        
+
         # Create corrupted file
         executions_path = tmp_path / "corrupt_test" / "executions"
         executions_path.mkdir(parents=True, exist_ok=True)
-        
+
         corrupt_file = executions_path / "corrupt.json"
         corrupt_file.write_text("{ invalid json }")
-        
+
         # Should handle gracefully
         loaded = await backend.load("corrupt")
         assert loaded is None
-        
+
         # List should skip corrupted files
         executions = await backend.list_executions()
         assert "corrupt" not in executions
@@ -205,34 +199,39 @@ class TestAdaptiveMemory:
         """Test learning from successful execution"""
         backend = InMemoryBackend()
         adaptive = AdaptiveMemory(backend)
-        
+
         # Create execution with results
         memory = ExecutionMemory(
             execution_id="learn-test",
             objective="Research AI applications in healthcare",
             task_type=TaskType.RESEARCH,
-            research_history=[["Found medical AI uses"], ["Identified key applications"]],
+            research_history=[
+                ["Found medical AI uses"],
+                ["Identified key applications"],
+            ],
             subagent_results=[
                 SubagentResult(
                     aspect_name="Medical Imaging",
                     findings="AI used for diagnosis",
                     success=True,
-                    start_time=datetime.now()
+                    start_time=datetime.now(),
                 ),
                 SubagentResult(
                     aspect_name="Drug Discovery",
                     findings="AI accelerates research",
                     success=True,
-                    start_time=datetime.now()
-                )
-            ]
+                    start_time=datetime.now(),
+                ),
+            ],
         )
-        
+
         # Learn from execution
         await adaptive.learn_from_execution(memory)
-        
+
         # Check session patterns
-        research_patterns = adaptive.session_patterns["successful_approaches"]["research"]
+        research_patterns = adaptive.session_patterns["successful_approaches"][
+            "research"
+        ]
         assert len(research_patterns) == 2  # One for each iteration
 
     @pytest.mark.asyncio
@@ -240,7 +239,7 @@ class TestAdaptiveMemory:
         """Test approach suggestion based on similarity"""
         backend = InMemoryBackend()
         adaptive = AdaptiveMemory(backend)
-        
+
         # Add some patterns
         adaptive.session_patterns["successful_approaches"]["research"] = [
             {
@@ -248,33 +247,31 @@ class TestAdaptiveMemory:
                 "objective_snippet": "Research quantum computing applications",
                 "successful_aspects": ["Quantum algorithms", "Hardware"],
                 "synthesis": "Found key applications",
-                "timestamp": datetime.now().isoformat()
+                "timestamp": datetime.now().isoformat(),
             },
             {
                 "task_type": "research",
                 "objective_snippet": "Analyze machine learning frameworks",
                 "successful_aspects": ["TensorFlow", "PyTorch"],
                 "synthesis": "Compared frameworks",
-                "timestamp": datetime.now().isoformat()
-            }
+                "timestamp": datetime.now().isoformat(),
+            },
         ]
-        
+
         # Test similar objective
         suggestion = await adaptive.suggest_approach(
-            "Research quantum computing algorithms",
-            TaskType.RESEARCH
+            "Research quantum computing algorithms", TaskType.RESEARCH
         )
-        
+
         assert suggestion is not None
         assert "Quantum algorithms" in suggestion["suggested_aspects"]
         assert suggestion["confidence"] > 0
 
         # Test dissimilar objective
         suggestion = await adaptive.suggest_approach(
-            "Fix the bug in the code",
-            TaskType.RESEARCH
+            "Fix the bug in the code", TaskType.RESEARCH
         )
-        
+
         # Should not find good match
         assert suggestion is None or suggestion["confidence"] < 0.3
 
@@ -282,16 +279,16 @@ class TestAdaptiveMemory:
         """Test getting effective tools for task type"""
         backend = InMemoryBackend()
         adaptive = AdaptiveMemory(backend)
-        
+
         # Add tool effectiveness data
         adaptive.session_patterns["tool_effectiveness"]["research"] = {
             "web_search": 5,
             "filesystem": 2,
-            "database": 1
+            "database": 1,
         }
-        
+
         effective_tools = adaptive.get_effective_tools(TaskType.RESEARCH)
-        
+
         # Should be sorted by effectiveness
         assert effective_tools[0] == "web_search"
         assert effective_tools[1] == "filesystem"
@@ -305,7 +302,7 @@ class TestMemoryManager:
     async def test_memory_manager_with_learning(self):
         """Test memory manager with learning enabled"""
         manager = MemoryManager(enable_learning=True)
-        
+
         # Create and save memory
         memory = ExecutionMemory(
             execution_id="mgr-test",
@@ -316,19 +313,18 @@ class TestMemoryManager:
                     aspect_name="File Edit",
                     findings="Successfully edited",
                     success=True,
-                    start_time=datetime.now()
+                    start_time=datetime.now(),
                 )
-            ]
+            ],
         )
-        
+
         await manager.save_memory(memory)
-        
+
         # Should trigger learning
         suggestion = await manager.suggest_approach(
-            "Edit another file",
-            TaskType.ACTION
+            "Edit another file", TaskType.ACTION
         )
-        
+
         # May or may not have suggestions depending on similarity
         assert suggestion is None or isinstance(suggestion, dict)
 
@@ -336,19 +332,19 @@ class TestMemoryManager:
     async def test_memory_manager_without_learning(self):
         """Test memory manager with learning disabled"""
         manager = MemoryManager(enable_learning=False)
-        
+
         memory = ExecutionMemory(
             execution_id="no-learn",
             objective="Test without learning",
-            task_type=TaskType.RESEARCH
+            task_type=TaskType.RESEARCH,
         )
-        
+
         await manager.save_memory(memory)
-        
+
         # Should not have learning capabilities
         suggestion = await manager.suggest_approach("Any objective", TaskType.RESEARCH)
         assert suggestion is None
-        
+
         tools = manager.get_effective_tools(TaskType.RESEARCH)
         assert tools == []
 
@@ -356,20 +352,22 @@ class TestMemoryManager:
     async def test_list_executions_through_manager(self):
         """Test listing executions through manager"""
         manager = MemoryManager()
-        
+
         # Save multiple executions
         for i in range(3):
             memory = ExecutionMemory(
                 execution_id=f"list-{i}",
                 objective=f"Objective {i}",
                 task_type=TaskType.RESEARCH,
-                total_cost=i * 1.5
+                total_cost=i * 1.5,
             )
             await manager.save_memory(memory)
-        
+
         # List executions
         executions = await manager.list_executions()
-        
+
         assert len(executions) == 3
         assert executions["list-1"]["total_cost"] == 1.5
-        assert all(exec_data["task_type"] == "research" for exec_data in executions.values())
+        assert all(
+            exec_data["task_type"] == "research" for exec_data in executions.values()
+        )
